# docs/openapi.json
{
  "components": {
    "schemas": {
      "AddExerciseRequest": {
        "properties": {
          "exercise_id": {
            "minimum": 1.0,
            "title": "Exercise Id",
            "type": "integer"
          }
        },
        "required": [
          "exercise_id"
        ],
        "title": "AddExerciseRequest",
        "type": "object"
      },
      "AddSetRequest": {
        "properties": {
          "reps": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Reps",
            "type": "integer"
          },
          "weight": {
            "minimum": 0.0,
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "weight",
          "reps"
        ],
        "title": "AddSetRequest",
        "type": "object"
      },
...
      "CurrentWorkoutResponse": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "elapsed_seconds": {
            "title": "Elapsed Seconds",
            "type": "integer"
          },
          "exercises": {
            "items": {
              "$ref": "#/components/schemas/CurrentWorkoutExerciseResponse"
            },
            "title": "Exercises",
            "type": "array"
          },
          "load_metrics": {
            "$ref": "#/components/schemas/LoadMetricsResponse"
          },
          "lower_back_pain": {
            "title": "Lower Back Pain",
            "type": "integer"
          },
          "next_workout_recommendation": {
            "$ref": "#/components/schemas/NextWorkoutRecommendationResponse"
          },
          "recovery_context": {
            "$ref": "#/components/schemas/RecoveryContextResponse"
          },
          "session_rpe": {
            "title": "Session Rpe",
            "type": "integer"
          },
          "started_at": {
            "title": "Started At",
            "type": "string"
          },
          "total_reps": {
            "title": "Total Reps",
            "type": "integer"
          },
          "total_sets": {
            "title": "Total Sets",
            "type": "integer"
          },
          "total_volume": {
            "title": "Total Volume",
            "type": "number"
          }
        },
        "required": [
          "active",
          "elapsed_seconds",
          "total_sets",
          "total_reps",
          "total_volume",
          "exercises"
        ],
        "title": "CurrentWorkoutResponse",
        "type": "object"
      },
...
      "ExerciseMutationResponse": {
        "properties": {
          "created": {
            "title": "Created",
            "type": "boolean"
          },
          "exercise": {
            "$ref": "#/components/schemas/ExerciseResponse"
          }
        },
        "required": [
          "exercise"
        ],
        "title": "ExerciseMutationResponse",
        "type": "object"
      },
...
      "ExercisesResponse": {
        "properties": {
          "exercises": {
            "items": {
              "$ref": "#/components/schemas/ExerciseResponse"
            },
            "title": "Exercises",
            "type": "array"
          }
        },
        "required": [
          "exercises"
        ],
        "title": "ExercisesResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UpdateSetRequest": {
        "properties": {
          "reps": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Reps",
            "type": "integer"
          },
          "weight": {
            "minimum": 0.0,
            "title": "Weight",
            "type": "number"
          }
        },
        "title": "UpdateSetRequest",
        "type": "object"
      },
...
      "WorkoutMetadataUpdate": {
        "properties": {
          "lower_back_pain": {
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Lower Back Pain",
            "type": "integer"
          },
          "session_rpe": {
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Session Rpe",
            "type": "integer"
          }
        },
        "title": "WorkoutMetadataUpdate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Training Log",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/current-workout/exercises": {
      "post": {
        "operationId": "add_current_workout_exercise_api_v1_current_workout_exercises_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddExerciseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentWorkoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Current Workout Exercise",
        "tags": [
          "current-workout"
        ]
      }
    },
    "/api/v1/current-workout/exercises/{draft_exercise_id}": {
      "delete": {
        "operationId": "delete_current_workout_exercise_api_v1_current_workout_exercises__draft_exercise_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "draft_exercise_id",
            "required": true,
            "schema": {
              "title": "Draft Exercise Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentWorkoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Current Workout Exercise",
        "tags": [
          "current-workout"
        ]
      }
    },
    "/api/v1/current-workout/exercises/{draft_exercise_id}/sets": {
      "post": {
        "operationId": "add_current_workout_set_api_v1_current_workout_exercises__draft_exercise_id__sets_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_exercise_id",
            "required": true,
            "schema": {
              "title": "Draft Exercise Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentWorkoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Current Workout Set",
        "tags": [
          "current-workout"
        ]
      }
    },
    "/api/v1/current-workout/exercises/{draft_exercise_id}/sets/duplicate": {
      "post": {
        "operationId": "duplicate_current_workout_set_api_v1_current_workout_exercises__draft_exercise_id__sets_duplicate_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_exercise_id",
            "required": true,
            "schema": {
              "title": "Draft Exercise Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentWorkoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Duplicate Current Workout Set",
        "tags": [
          "current-workout"
        ]
      }
    },
...
    "/api/v1/current-workout/metadata": {
      "patch": {
        "operationId": "update_current_workout_metadata_api_v1_current_workout_metadata_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkoutMetadataUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentWorkoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Current Workout Metadata",
        "tags": [
          "current-workout"
        ]
      }
    },
    "/api/v1/current-workout/sets/{draft_set_id}": {
      "delete": {
        "operationId": "delete_current_workout_set_api_v1_current_workout_sets__draft_set_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "draft_set_id",
            "required": true,
            "schema": {
              "title": "Draft Set Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentWorkoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Current Workout Set",
        "tags": [
          "current-workout"
        ]
      },
      "patch": {
        "operationId": "update_current_workout_set_api_v1_current_workout_sets__draft_set_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "draft_set_id",
            "required": true,
            "schema": {
              "title": "Draft Set Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentWorkoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Current Workout Set",
        "tags": [
          "current-workout"
        ]
      }
    },
...
    "/api/v1/exercises": {
      "get": {
        "operationId": "get_exercises_api_v1_exercises_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_inactive",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Inactive",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExercisesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Exercises",
        "tags": [
          "exercises"
        ]
      },
      "post": {
        "operationId": "create_exercise_endpoint_api_v1_exercises_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Exercise Endpoint",
        "tags": [
          "exercises"
        ]
      }
    },
...
    "/api/v1/exercises/{exercise_id}": {
      "patch": {
        "operationId": "update_exercise_endpoint_api_v1_exercises__exercise_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "exercise_id",
            "required": true,
            "schema": {
              "title": "Exercise Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Exercise Endpoint",
        "tags": [
          "exercises"
        ]
      }
    },
...
    "/api/v1/sets/{set_id}": {
      "delete": {
        "operationId": "delete_set_endpoint_api_v1_sets__set_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "set_id",
            "required": true,
            "schema": {
              "title": "Set Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Set Endpoint",
        "tags": [
          "workouts"
        ]
      },
      "patch": {
        "operationId": "update_set_endpoint_api_v1_sets__set_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "set_id",
            "required": true,
            "schema": {
              "title": "Set Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Set Endpoint",
        "tags": [
          "workouts"
        ]
      }
    },
...
    "/api/v1/workouts": {
      "get": {
        "operationId": "get_workouts_api_v1_workouts_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 500.0,
              "minimum": 1.0,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workouts",
        "tags": [
          "workouts"
        ]
      }
    },
    "/api/v1/workouts/{workout_id}": {
      "delete": {
        "operationId": "delete_workout_endpoint_api_v1_workouts__workout_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "workout_id",
            "required": true,
            "schema": {
              "title": "Workout Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteWorkoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Workout Endpoint",
        "tags": [
          "workouts"
        ]
      },
      "get": {
        "operationId": "get_workout_detail_api_v1_workouts__workout_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workout_id",
            "required": true,
            "schema": {
              "title": "Workout Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workout Detail",
        "tags": [
          "workouts"
        ]
      },
      "patch": {
        "operationId": "update_workout_endpoint_api_v1_workouts__workout_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "workout_id",
            "required": true,
            "schema": {
              "title": "Workout Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkoutUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Workout Endpoint",
        "tags": [
          "workouts"
        ]
      }
    },
    "/api/v1/workouts/{workout_id}/exercises": {
      "post": {
        "operationId": "add_workout_exercise_endpoint_api_v1_workouts__workout_id__exercises_post",
        "parameters": [
          {
            "in": "path",
            "name": "workout_id",
            "required": true,
            "schema": {
              "title": "Workout Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddExerciseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Workout Exercise Endpoint",
        "tags": [
          "workouts"
        ]
      }
    },
    "/api/v1/workouts/{workout_id}/exercises/{workout_exercise_id}": {
      "delete": {
        "operationId": "delete_workout_exercise_endpoint_api_v1_workouts__workout_id__exercises__workout_exercise_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "workout_id",
            "required": true,
            "schema": {
              "title": "Workout Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "workout_exercise_id",
            "required": true,
            "schema": {
              "title": "Workout Exercise Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Workout Exercise Endpoint",
        "tags": [
          "workouts"
        ]
      }
    }
  }
}